feat: add logout functionality
Now, the user can log out. The way it works is with a table in DB that stores the tokens, as a whitelist of tokens. When a user logs out, the token he/she was using is deleted from that whitelist, and the access won't be possible again with that token.